Update writer for viem after each chan switch with updated `chain` object
