+++ cypress/lib/testname.js
@@ -1,6 +1,7 @@
 function testname() {
   return Cypress.currentTest.titlePath.join('--').replace(/[^\w-]+/gi, '-');
 }
+
 module.exports = {
   testname,
 };
